Function `sm_contains` is in a separate module.
