+++ syspathmodif/__init__.py
@@ -1,7 +1,9 @@
 from .syspathbundle import SysPathBundle
 
+from .sys_modules import\
+	sm_contains
+
 from .individual_paths import\
-	sm_contains,\
 	sp_append,\
 	sp_contains,\
 	sp_remove
@@ -9,7 +11,7 @@ from .individual_paths import\
 
 __all__ = [
 	SysPathBundle.__name__,
-	sm_contains,
+	sm_contains.__name__,
 	sp_append.__name__,
 	sp_contains.__name__,
 	sp_remove.__name__
