+++ syspathmodif/individual_paths.py
@@ -9,21 +9,6 @@ from .individual_paths_no_type_check import\
 	sp_remove_no_type_check
 
 
-def sm_contains(module_name):
-	"""
-	Indicates whether dictionary sys.modules contains the module or package
-	whose name is the argument.
-
-	Args:
-		module_name (str): the name of a module or package.
-
-	Returns:
-		bool: True if sys.modules contains argument module_name, False
-			otherwise.
-	"""
-	return module_name in sys.modules
-
-
 def sp_append(some_path):
 	"""
 	Appends the given path to the end of list sys.path if it does not already
@@ -81,7 +66,6 @@ def sp_remove(some_path):
 
 
 __all__ = [
-	sm_contains.__name__,
 	sp_append.__name__,
 	sp_contains.__name__,
 	sp_remove.__name__
